refactor: modularize Crocodile
